refactor/ editor,highlight,pdf 스토어를 한 폴더로 모음
