+++ src/store/file/editor.ts
