+++ src/store/file/highlight.ts
